QSIG/requirements.txt
bgraph
sh
typer
yara-python
editdistance
python-magic
firmwareextractor
quokka-project
python-json-logger